@@ -19,14 +19,14 @@
         }
     ],
     "require": {
-        "php": "^7.2",
+        "php": "^7.4",
         "ext-pdo": "*",
         "squirrelphp/debug": "^0.5",
         "doctrine/dbal": "^2.5"
     },
     "require-dev": {
         "bamarni/composer-bin-plugin": "^1.3",
-        "captainhook/plugin-composer": "^4.0",
+        "captainhook/plugin-composer": "^5.0",
         "mockery/mockery": "^1.0"
     },
     "suggest": {
@@ -37,7 +37,7 @@
     "config": {
         "sort-packages": false,
         "platform": {
-            "php": "7.2"
+            "php": "7.4"
         }
     },
     "autoload": {
@@ -51,16 +51,17 @@
         }
     },
     "scripts": {
-        "post-install-cmd": ["@composer bin all install --ansi"],
-        "post-update-cmd": ["@composer bin all update --ansi"],
         "phpstan": "vendor/bin/phpstan analyse",
-        "phpstan_base": "vendor/bin/phpstan analyse --configuration phpstan_base.neon --error-format baselineNeon > phpstan-baseline.neon",
-        "psalm": "vendor/bin/psalm --show-info=false",
+        "phpstan_base": "vendor/bin/phpstan analyse --generate-baseline",
+        "phpstan_clear": "vendor/bin/phpstan clear-result-cache",
+        "psalm": "vendor/bin/psalm --show-info=false --diff",
+        "psalm_full": "vendor/bin/psalm --show-info=false",
         "psalm_base": "vendor/bin/psalm --set-baseline=psalm-baseline.xml",
         "phpunit": "vendor/bin/phpunit --colors=always",
-        "phpcs": "vendor/bin/phpcs --standard=ruleset.xml --extensions=php --cache src tests",
-        "phpcsfix": "vendor/bin/phpcbf --standard=ruleset.xml --extensions=php --cache src tests",
+        "phpunit_clover": "vendor/bin/phpunit --coverage-text --coverage-clover build/logs/clover.xml",
         "codecoverage": "vendor/bin/phpunit --coverage-html tests/_reports",
+        "phpcs": "vendor/bin/phpcs --standard=ruleset.xml --extensions=php --cache=.phpcs-cache --colors src tests",
+        "phpcsfix": "vendor/bin/phpcbf --standard=ruleset.xml --extensions=php --cache=.phpcs-cache src tests",
         "binupdate": "@composer bin all update --ansi",
         "bininstall": "@composer bin all install --ansi"
     }
