@@ -11,50 +11,47 @@ use Squirrel\Queries\DBInterface;
  */
 class SelectEntries implements \IteratorAggregate
 {
-    /**
-     * @var DBInterface
-     */
-    private $db;
+    private DBInterface $db;
 
     /**
      * @var array<int|string,string> Only retrieve these fields of the tables
      */
-    private $fields = [];
+    private array $fields = [];
 
     /**
      * @var array<int|string,mixed>
      */
-    private $tables = [];
+    private array $tables = [];
 
     /**
      * @var array<int|string,mixed> WHERE restrictions in query
      */
-    private $where = [];
+    private array $where = [];
 
     /**
      * @var array<int|string,string> ORDER BY sorting in query
      */
-    private $orderBy = [];
+    private array $orderBy = [];
 
     /**
      * @var array<int|string,string> GROUP BY aggregating in query
      */
-    private $groupBy = [];
+    private array $groupBy = [];
 
     /**
      * @var int How many results should be returned
      */
-    private $limitTo = 0;
+    private int $limitTo = 0;
 
     /**
      * @var int Where in the result set to start (so many entries are skipped)
      */
-    private $startAt = 0;
+    private int $startAt = 0;
 
     /**
      * @var bool Whether the SELECT query should block the scanned entries
      */
-    private $blocking = false;
+    private bool $blocking = false;
 
     public function __construct(DBInterface $db)
     {
@@ -101,7 +98,6 @@ class SelectEntries implements \IteratorAggregate
 
     /**
      * @param array<int|string,string>|string $orderByClauses
-     * @return SelectEntries
      */
     public function orderBy($orderByClauses): self
     {
@@ -115,7 +111,6 @@ class SelectEntries implements \IteratorAggregate
 
     /**
      * @param array<int|string,string>|string $groupByClauses
-     * @return SelectEntries
      */
     public function groupBy($groupByClauses): self
     {
@@ -183,7 +178,7 @@ class SelectEntries implements \IteratorAggregate
      */
     public function getFlattenedFields(): array
     {
-        return $this->db->fetchAll([
+        return $this->db->fetchAllAndFlatten([
             'fields' => $this->fields,
             'tables' => $this->tables,
             'where' => $this->where,
@@ -192,14 +187,10 @@ class SelectEntries implements \IteratorAggregate
             'limit' => $this->limitTo,
             'offset' => $this->startAt,
             'lock' => $this->blocking,
-            'flattenFields' => true,
         ]);
     }
 
-    /**
-     * @return SelectIterator
-     */
-    public function getIterator()
+    public function getIterator(): SelectIterator
     {
         return new SelectIterator($this->db, [
             'fields' => $this->fields,
