@@ -7,22 +7,13 @@ use Squirrel\Queries\DBSelectQueryInterface;
 
 class DBSelectQuery implements DBSelectQueryInterface
 {
-    /**
-     * @var ResultStatement
-     */
-    private $statement;
+    private ResultStatement $statement;
 
-    /**
-     * @param ResultStatement $statement
-     */
     public function __construct(ResultStatement $statement)
     {
         $this->statement = $statement;
     }
 
-    /**
-     * @return ResultStatement
-     */
     public function getStatement(): ResultStatement
     {
         return $this->statement;
